kids-story-automation/requirements.txt
edge-tts
moviepy
pillow
flask
google-auth-oauthlib
google-api-python-client
requests
